Add dotnet cli to runtime test Helix jobs
Currently managed tools like crossgen2 are run against the test runtime which is slow on Debug builds, and can obscure errors when the whole XUnit infra fails due to an introduced runtime bug.

Added a `xunit.console.runtimeconfig.dev.json` which forces allows `xunit.console` to use the repo-local dotnet.cmd on dev boxes and the same version installed on the path in Helix.

When running crossgen2 during test execution, support both local dev and Helix scenario when deciding which dotnet to run. On Helix, simply use `dotnet` which assumes a compatible dotnet runtime is in the path. Locally, tests are run with runtest.cmd|sh which sets `__TestDotNetCmd` to the repo-local dotnet script. This preserves the characteristic that no machine-wide 5.0 dotnet runtime must be installed for the runtime tests.

Our batch scripting needs to also use `call dotnet.cmd` when running Crossgen2, whereas previously it directly executed corerun.exe.