Attempt to fix CG2 VS build for test assemblies
ILTestAssembly and CoreTestAssembly are not getting built correctly in, ie, checked mode builds of CG2 in VS. The configuration / platform are not passed through to the P2P project build correctly due to some special handling around projects that are referenced by solutions but not part of the solution build. They're built with the project's default config causing the later copy step to fail to find the assembly.

Corefx fixed a similar issue via https://github.com/dotnet/runtime/pull/32259.

This needs more testing. From a totally clean build, this still is insufficient to get the test assemblies built.